Sed TYPE to ENGINE in created SQL queries.
As of MySQL 5.1.8, TYPE = engine_name is still accepted as a synonym
for the ENGINE = engine_name table option but generates a warning. You
should note that this option is not available in MySQL 5.1.7, and is
removed altogether in MySQL 5.5 and produces a syntax error.
TYPE has been deprecated since MySQL 4.0.